+++ app/src/main/java/com/example/myapplication1/ContactMessages.kt
@@ -81,6 +81,7 @@ class ContactMessages : Fragment() {
         return view
     }
 
+    @RequiresApi(Build.VERSION_CODES.O)
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
